Consolidate chat handlers with base class pattern
Refactor to eliminate ~95% code duplication between container and host
chat handlers by extracting shared logic into abstract base classes:

- Create src/chat/types.ts with shared interfaces (ChatMessage,
  ClaudeStreamMessage, OpencodeStreamEvent, SpawnConfig, etc.)
- Create src/chat/base-claude-session.ts with shared Claude Code logic
  (buffer processing, stream message handling, interrupt, setModel)
- Create src/chat/base-opencode-session.ts with shared OpenCode logic
  (buffer processing, stream event handling, interrupt, setModel)
- Refactor ChatSession and HostChatSession to extend BaseClaudeSession
- Refactor OpencodeSession and HostOpencodeSession to extend
  BaseOpencodeSession
- Add DEFAULT_CLAUDE_MODEL constant to src/shared/constants.ts

Each handler now only implements:
- getSpawnConfig(): command and spawn options specific to container/host
- getLogPrefix(): logging identifier
- getNoOutputErrorMessage(): context-specific error message
- loadHistory() (OpenCode only): history loading strategy

This reduces code from ~850 lines across 4 files to ~250 lines in
handlers plus ~350 lines in shared base classes.
